backend/requirements.txt
# Production-grade Echo API requirements
# Install with: pip install -r requirements.txt

# Core FastAPI and async support
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
psutil==5.9.6

# Database and ORM
sqlalchemy==2.0.23
sqlmodel==0.0.14
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# Caching and message broker
redis==5.0.1
fakeredis==2.21.1
celery==5.3.4

# AI and ML
openai>=1.40.0
chromadb==0.4.18
numpy>=1.26.0
scikit-learn>=1.3.2

# Authentication and security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
bleach==6.1.0
jwcrypto==1.5.6
authlib==1.3.0
pyotp==2.9.0
qrcode==7.4.2
aiosmtplib==3.0.1

# Configuration and environment
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
jinja2==3.1.2

# WebSocket
websockets==12.0

# Media Storage (MinIO - S3 compatible)
minio==7.2.0
boto3==1.29.7

# Payment Processing
stripe==7.6.0

# Additional ML/AI
sentence-transformers==2.2.2

# Monitoring
sentry-sdk[fastapi]==1.38.0
aiofiles==23.2.1
